feat: Novas funcionalidades estruturais

+++ routes/web.php
@@ -8,9 +8,23 @@
 use Glauce\Router\Router;
 
 Router::get('/', 'HomeController@index');
-Router::get('/user', 'UserController@index');
-Router::get('/user/create', 'UserController@create');
-Router::post('/user/create', 'UserController@create');
-Router::get('/user/update/{id}', 'UserController@update');
-Router::put('/user/update/{id}', 'UserController@update');
-Router::delete('/user/delete/{id}', 'UserController@delete');
+Router::get('/about', 'HomeController@about');
+Router::get('/works', 'HomeController@works');
+Router::get('/contact', 'HomeController@contact');
+
+Router::get('/login', 'AuthController@login');
+Router::get('/register', 'AuthController@register');
+Router::post('/auth', 'AuthController@loginEnter');
+Router::post('/register', 'AuthController@registerEnter');
+Router::get('/logout', 'AuthController@logout');
+
+Router::get('/perfil', 'UserController@perfil');
+
+Router::get('/admin', 'Admin\HomeController@index');
+
+Router::get('/admin/user', 'UserController@index', ['AuthMiddleware']);
+Router::get('/admin/user/create', 'UserController@create');
+Router::post('/admin/user/insert', 'UserController@insert');
+Router::get('/admin/user/update/{id}', 'UserController@update');
+Router::put('/admin/user/update/{id}', 'UserController@update');
+Router::delete('/admin/user/delete/{id}', 'UserController@delete');
